# Tweaking main scripts and lib