--- a/ThumbyGames/lib-emulator/thumby.py
+++ b/ThumbyGames/lib-emulator/thumby.py
@@ -1,9 +1,10 @@
 # Thumby module.
+# - Emulator edition
 
 # Contains helpful abstractions between hardware features of Thumby and the uPython REPL.
 
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -21,20 +22,16 @@
     the Thumby API. If not, see <https://www.gnu.org/licenses/>.
 '''
 
-# Necessary things.
-#from machine import Pin, Timer, I2C, PWM, SPI, UART
-from machine import reset as machineReset
+from machine import reset as machineReset #, freq
 import emulator
-#import ssd1306
-#import os
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 # Grab initial frequency
 # __f0 = freq()
 # Speed us up so imports take less time
-# freq(250000000)
+# freq(250_000_000)
 
 from thumbyHardware import swL, swR, swU, swD, swA, swB, swBuzzer, IDPin, i2c, spi