--- a/ThumbyGames/lib-emulator/thumbyGraphics.py
+++ b/ThumbyGames/lib-emulator/thumbyGraphics.py
@@ -1,6 +1,8 @@
 # Thumby graphics base
+# - Emulator edition
+
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -26,7 +28,7 @@ from thumbyHardware import i2c, spi
 from thumbyButton import buttonA, buttonB, buttonU, buttonD, buttonL, buttonR
 import emulator
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 # Graphics class, from which the gfx namespace is defined.