--- a/ThumbyGames/lib-emulator/thumbyHardware.py
+++ b/ThumbyGames/lib-emulator/thumbyHardware.py
@@ -20,11 +20,11 @@
     the Thumby API. If not, see <https://www.gnu.org/licenses/>.
 '''
 
-from machine import Pin, Timer, I2C, PWM, SPI
+from machine import Pin, PWM, SPI
 from machine import reset as machineReset
 import emulator
 
-# Last updated 14-Dec-2022
+# Last updated 15-Dec-2022
 __version__ = '1.9'
 
 # Pin definitions for button inputs & buzzer.
@@ -54,12 +54,15 @@ IDPin = Pin(14, Pin.IN, Pin.PULL_DOWN)
 IDPin = Pin(13, Pin.IN, Pin.PULL_DOWN)
 IDPin = Pin(12, Pin.IN, Pin.PULL_DOWN)
 
-i2c = None
-spi = None
-if(HWID==0):
-    i2c = I2C(0, sda=Pin(16), scl=Pin(17), freq=1000000)
+
 if(HWID>=1):
-    spi = SPI(0, sck=Pin(18), mosi=Pin(19))#possible assignment of miso to 4 or 16?
+    spi = SPI(0, sck=Pin(18), mosi=Pin(19)) # Assign miso to 4 or 16?
+    i2c = None
+else:
+    from machine import I2C
+    i2c = I2C(0, sda=Pin(16), scl=Pin(17), freq=1_000_000)
+    spi = None
+
 
 # Wrap machine.reset() to be accessible as thumby.reset()
 def reset():