--- a/ThumbyGames/lib-emulator/thumbyLink.py
+++ b/ThumbyGames/lib-emulator/thumbyLink.py
@@ -1,6 +1,8 @@
 # Thumby link base
+# - Emulator edition
+
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -22,7 +24,7 @@ from machine import Pin, UART
 from time import ticks_ms, ticks_diff
 import emulator
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 class LinkClass: