--- a/ThumbyGames/lib-emulator/thumbySaves.py
+++ b/ThumbyGames/lib-emulator/thumbySaves.py
@@ -22,46 +22,46 @@
 
 from json import load as JSONLoad, dump as JSONDump
 from ubinascii import a2b_base64 as b64dec, b2a_base64 as b64enc
-import os
+from os import chdir, getcwd, mkdir, rename, remove, stat
 import emulator
 
-# Last updated 14-Dec-2022
+# Last updated 20-Dec-2022
 __version__ = '1.9'
 
 class SavesClass:
     def __init__(self):
         
-        oldDir = os.getcwd()
+        oldDir = getcwd()
         try:
-            os.stat("/Saves")
+            stat("/Saves")
         except OSError:
-            os.chdir("/")
-            os.mkdir("/Saves")
-            os.mkdir("/Saves/temp")
+            chdir("/")
+            mkdir("/Saves")
+            mkdir("/Saves/temp")
             
         self.savesPath = "/Saves"
         self.saveFile = None
         self.volatileDict = dict()
         
         try:
-            os.stat("/Saves/temp")
+            stat("/Saves/temp")
             self.setName("temp")
         except OSError:
             pass
         
-        os.chdir(oldDir)
+        chdir(oldDir)
     
     # Set a game save's working subdirectory in "/Saves/"
     # @micropython.viper
     def setName(self, subdir):
         
-        oldDir = os.getcwd()
+        oldDir = getcwd()
         self.savesPath = "/Saves/" + subdir
         
         try:
-            os.stat(self.savesPath)
+            stat(self.savesPath)
         except OSError:
-            os.mkdir(self.savesPath)
+            mkdir(self.savesPath)
         
         if type(self.saveFile) != type(None):
             self.saveFile.close()
@@ -79,7 +79,7 @@ class SavesClass:
                 self.saveFile.write("{}")
                 self.saveFile.seek(0, 0)
                 self.volatileDict = JSONLoad(self.saveFile)
-        os.chdir(oldDir)
+        chdir(oldDir)
     
     # Set entry in volatile dictionary
     @micropython.native
@@ -127,7 +127,7 @@ class SavesClass:
     # @micropython.native
     def save(self, backup = False):
         
-        oldDir = os.getcwd()
+        oldDir = getcwd()
         
         if(self.savesPath == "/Saves"): # If a directory hasn't been set, use a temporary one
             self.savesPath = "/Saves/temp"
@@ -137,16 +137,16 @@ class SavesClass:
         
         try:
             if(backup == True):
-                os.rename(self.savesPath+"/persistent.json", self.savesPath+"/backup.json")
+                rename(self.savesPath+"/persistent.json", self.savesPath+"/backup.json")
             else:
-                os.remove(self.savesPath+"/persistent.json")
+                remove(self.savesPath+"/persistent.json")
         except OSError:
             pass
         
         self.saveFile = open(self.savesPath+"/persistent.json", "w+")
         JSONDump(self.volatileDict, self.saveFile)
         self.saveFile.flush()
-        os.chdir(oldDir)
+        chdir(oldDir)
     
     # Return the current save path
     @micropython.viper