--- a/ThumbyGames/lib/thumby.py
+++ b/ThumbyGames/lib/thumby.py
@@ -3,7 +3,7 @@
 # Contains helpful abstractions between hardware features of Thumby and the uPython REPL.
 
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -21,11 +21,11 @@
     the Thumby API. If not, see <https://www.gnu.org/licenses/>.
 '''
 
+from machine import reset as machineReset, freq
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
-from machine import reset as machineReset, freq
 # Grab initial frequency
 __f0 = freq()
 # Speed us up so imports take less time