--- a/ThumbyGames/lib/thumbyAudio.py
+++ b/ThumbyGames/lib/thumbyAudio.py
@@ -1,6 +1,7 @@
 # Thumby audio base
+
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -22,7 +23,7 @@ from thumbyHardware import swBuzzer
 from machine import Timer
 from time import ticks_ms, ticks_diff
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 # Audio class, from which the audio namespace is defined.