--- a/ThumbyGames/lib/thumbyGraphics.py
+++ b/ThumbyGames/lib/thumbyGraphics.py
@@ -1,6 +1,7 @@
 # Thumby graphics base
+
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -25,7 +26,7 @@ from time import ticks_ms, ticks_diff, sleep_ms
 from thumbyHardware import i2c, spi
 from thumbyButton import buttonA, buttonB, buttonU, buttonD, buttonL, buttonR
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 # Graphics class, from which the gfx namespace is defined.