--- a/ThumbyGames/lib/thumbyHardware.py
+++ b/ThumbyGames/lib/thumbyHardware.py
@@ -19,10 +19,9 @@
     the Thumby API. If not, see <https://www.gnu.org/licenses/>.
 '''
 
-from machine import Pin, PWM, SPI
-from machine import reset as machineReset
+from machine import Pin, PWM, SPI, reset as machineReset
 
-# Last updated 22-Dec-2022
+# Last updated 27-Dec-2022
 __version__ = '1.9'
 
 # Pin definitions for button inputs & buzzer.
@@ -65,4 +64,3 @@ else:
 # Wrap machine.reset() to be accessible as thumby.reset()
 def reset():
     machineReset()
-