--- a/ThumbyGames/lib/thumbyLink.py
+++ b/ThumbyGames/lib/thumbyLink.py
@@ -1,6 +1,7 @@
 # Thumby link base
+
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -21,14 +22,14 @@
 from machine import Pin, UART
 from time import ticks_ms, ticks_diff
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 class LinkClass:
     def __init__(self):
         self.initialized = False
-    
-    
+
+
     @micropython.native
     def init(self):
         self.rxPin = Pin(1, Pin.IN)