--- a/ThumbyGames/lib/thumbySprite.py
+++ b/ThumbyGames/lib/thumbySprite.py
@@ -1,6 +1,7 @@
 # Thumby sprite base
+
 # Written by Mason Watmough, Jason Marcum, and Ben Rose for TinyCircuits.
-# Last edited 7/11/2022
+# 11-Jul-2022
 
 '''
     This file is part of the Thumby API.
@@ -20,7 +21,7 @@
 
 from os import stat
 
-# Last updated 11/11/2022 for menu reset change
+# Last updated 14-Dec-2022
 __version__ = '1.9'
 
 # Sprite class for holding pixel data