… CHANGELOG.txt
… js/play.js