The `main.py` and `menu.py` should be lean. 🗜💪

Small tweaks to clean things up. Mainly trimming unused imports.
Tested and looks ok. (Thanks @fuglaro for extensive review and testing).

### Speed
Benchmarked with `ticks_ms`:  
`main.py` boot reduced from ~754ms to ~246ms (now takes ~1/3rd of the time!). 
`menu.py` up to main loop, reduced from ~1766ms to ~1581ms.

### Memory
Memory savings using `print(gc.mem_free())`:
**Vanilla**
Before GC: mem_free=81376
After GC: mem_free=122112
**This PR** 
Before GC: mem_free=84288  (+2,912 bytes free) ~2.8kb!!
After GC: mem_free=122144  (+32 bytes free) - the GC was managing to remove all the unused imports pretty well.