@fuglaro 2022-12-21
Note for testing:
I think only three games on the Arcade currently use the thumby saves API:
AstroJumper
BurnsCharade
PicoPac