@masonova1 2023-01-09
> @masonova1 - Thanks for taking the time to review this PR (and many commits!)... Readability is perhaps subjective. May I ask you: Is there a need to communicate to the casual user, that the base address is at `0x40058000` and the first register '0' is at `0x0C`? I kinda think that's great for a data sheet, but practically, I want to just glance at this 'magic' address where we set and read a flag ...IMHO. I also think with using the next 3 registers, it starts to look cluttered. (See [here](http://forum.tinycircuits.com/index.php?topic=2517.0) and [here](https://github.com/ace-dent/TinyCircuits-Thumby-Code-Editor/blob/c3576ac4332a0f4e9dc069d13f6630fd79e236bf/ThumbyGames/main.py#L33-L79)). If you feel strongly on this, I can add back in the extra variables. Many thanks, Andrew

I'm honestly fine with less words. If somebody wants the exposition, they would need to go back to the definition to see what it means anyways. If somebody complains, we'll bring it back. We'll get this merged soon.